% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptex.R
\name{ptex_cache_info}
\alias{ptex_cache_info}
\title{Inspect Cache Usage}
\usage{
ptex_cache_info(cache)
}
\arguments{
\item{cache}{A cache from \code{\link[=ptex_cache]{ptex_cache()}}.}
}
\value{
A named numeric vector of memory bytes, file counts and block reads.
}
\description{
Inspect Cache Usage
}
