% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptex.R
\name{ptex_close}
\alias{ptex_close}
\title{Close a Texture or Cache}
\usage{
ptex_close(handle)
}
\arguments{
\item{handle}{A texture or cache handle. Closing an already closed handle is safe.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Close a Texture or Cache
}
