% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptex.R
\name{ptex_version}
\alias{ptex_version}
\title{Ptex Library Version}
\usage{
ptex_version()
}
\value{
The bundled upstream Ptex version as a character string.
}
\description{
Ptex Library Version
}
\examples{
ptex_version()
}
